Clear some lint warnings

+++ beecom_server/beecom_server/src/bp_packet/BPPacketRPRTACK.java
@@ -28,7 +28,7 @@ public class BPPacketRPRTACK extends BPPacket {
 	@Override
 	public boolean assembleVariableHeader() throws BPAssembleVrbHeaderException {
 		super.assembleVariableHeader();
-		byte flags = (byte)getVrbHead().getFlags();
+		byte flags = getVrbHead().getFlags();
 		getIoBuffer().put(flags);
 		int packSeq = (byte)getVrbHead().getPackSeq();
 		getIoBuffer().putUnsignedShort(packSeq);
